refactor(team building): `ProjectUtil`의 책임 분리
