+++ src/main/java/com/skhu/gdgocteambuildingproject/teambuilding/model/ProjectUtil.java
@@ -1,12 +1,10 @@
 package com.skhu.gdgocteambuildingproject.teambuilding.model;
 
-import static com.skhu.gdgocteambuildingproject.global.exception.ExceptionMessage.NOT_PARTICIPATED;
 import static com.skhu.gdgocteambuildingproject.global.exception.ExceptionMessage.PROJECT_NOT_EXIST;
 
 import com.skhu.gdgocteambuildingproject.teambuilding.domain.ProjectSchedule;
 import com.skhu.gdgocteambuildingproject.teambuilding.domain.TeamBuildingProject;
 import com.skhu.gdgocteambuildingproject.teambuilding.domain.enumtype.ScheduleType;
-import com.skhu.gdgocteambuildingproject.teambuilding.repository.ProjectParticipantRepository;
 import com.skhu.gdgocteambuildingproject.teambuilding.repository.TeamBuildingProjectRepository;
 import jakarta.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
@@ -20,20 +18,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class ProjectUtil {
 
-    private final ProjectParticipantRepository participantRepository;
     private final TeamBuildingProjectRepository projectRepository;
-
-    public void validateParticipation(long userId, long projectId) {
-        boolean participated = participantRepository.existsByUserIdAndProjectId(userId, projectId);
-
-        if (!participated) {
-            throw new IllegalStateException(NOT_PARTICIPATED.getMessage());
-        }
-    }
-
-    public boolean isParticipated(long userId, long projectId) {
-        return participantRepository.existsByUserIdAndProjectId(userId, projectId);
-    }
 
     public Optional<TeamBuildingProject> findCurrentProject() {
         List<TeamBuildingProject> unfinishedProjects = findUnfinishedProjects();
